[WFCORE-766] Move cli extension tests from manual mode into standalone
