--- testsuite/manualmode/src/test/resources/org/jboss/as/test/manualmode/management/cli/extensions/org.jboss.as.cli.CommandHandlerProvider
@@ -1,1 +1,0 @@
-org.jboss.as.test.manualmode.management.cli.extensions.CliExtCommandHandlerProvider
