+++ testsuite/standalone/src/test/java/org/jboss/as/test/integration/management/cli/extensions/DuplicateExtCommandTestCase.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.manualmode.management.cli.extensions;
+package org.jboss.as.test.integration.management.cli.extensions;
 
 import org.jboss.as.cli.CommandHandlerProvider;
 import org.jboss.as.controller.Extension;
@@ -41,10 +41,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.wildfly.core.testrunner.ManagementClient;
-import org.wildfly.core.testrunner.ServerControl;
-import org.wildfly.core.testrunner.ServerController;
 import org.wildfly.core.testrunner.WildflyTestRunner;
-import org.xnio.IoUtils;
 
 import javax.inject.Inject;
 import java.io.*;
@@ -56,13 +53,16 @@ import static org.junit.Assert.assertTrue;
  * @author Petr Kremensky pkremens@redhat.com
  */
 @RunWith(WildflyTestRunner.class)
-@ServerControl(manual = true)
 public class DuplicateExtCommandTestCase extends CliScriptTestBase {
 
     private static final String MODULE_NAME = "test.cli.extension.commands";
 
     @Inject
-    private static ServerController containerController;
+    protected static ManagementClient client;
+
+    @Inject
+    protected static ModelControllerClient controllerClient;
+
 
     private static TestModule testModule;
 
@@ -74,26 +74,21 @@ public class DuplicateExtCommandTestCase extends CliScriptTestBase {
 
     @AfterClass
     public static void tearDownServer() throws Exception {
-        ModelControllerClient client = null;
         try {
-            client = containerController.getClient().getControllerClient();
-            ModelNode subsystemResult = client.execute(Util.createRemoveOperation(PathAddress.pathAddress(CliExtCommandsSubsystemResourceDescription.PATH)));
-            ModelNode extensionResult = client.execute(Util.createRemoveOperation(PathAddress.pathAddress(ModelDescriptionConstants.EXTENSION, MODULE_NAME)));
+            ModelNode subsystemResult = controllerClient.execute(Util.createRemoveOperation(PathAddress.pathAddress(CliExtCommandsSubsystemResourceDescription.PATH)));
+            ModelNode extensionResult = controllerClient.execute(Util.createRemoveOperation(PathAddress.pathAddress(ModelDescriptionConstants.EXTENSION, MODULE_NAME)));
             ModelTestUtils.checkOutcome(subsystemResult);
             ModelTestUtils.checkOutcome(extensionResult);
         } finally {
-            containerController.stop();
             testModule.remove();
-            IoUtils.safeClose(client);
         }
-        containerController.stop();
     }
 
     @Test
     public void testExtensionCommandCollision() throws Exception {
         assertEquals(0,
-                execute(containerController.getClient().getMgmtAddress(),
-                        containerController.getClient().getMgmtPort(),
+                execute(client.getMgmtAddress(),
+                        client.getMgmtPort(),
                         true,
                         "extension-commands --errors",
                         true));
@@ -134,15 +129,11 @@ public class DuplicateExtCommandTestCase extends CliScriptTestBase {
     }
 
     private static void setupServerWithExtension() throws Exception {
-        containerController.start();
-        ManagementClient managementClient = containerController.getClient();
-        ModelControllerClient client = managementClient.getControllerClient();
-
         //Add the extension
         final ModelNode addExtension = Util.createAddOperation(PathAddress.pathAddress(ModelDescriptionConstants.EXTENSION, MODULE_NAME));
-        ModelTestUtils.checkOutcome(client.execute(addExtension));
+        ModelTestUtils.checkOutcome(controllerClient.execute(addExtension));
 
         final ModelNode addSubsystem = Util.createAddOperation(PathAddress.pathAddress(CliExtCommandsSubsystemResourceDescription.PATH));
-        ModelTestUtils.checkOutcome(client.execute(addSubsystem));
+        ModelTestUtils.checkOutcome(controllerClient.execute(addSubsystem));
     }
 }
