+++ testsuite/standalone/src/test/resources/org/jboss/as/test/integration/management/cli/extensions/DuplicateExtCommandTestCase-module.xml
