+++ testsuite/standalone/src/test/resources/org/jboss/as/test/integration/management/cli/extensions/org.jboss.as.controller.Extension
@@ -1,0 +1,1 @@
+org.jboss.as.test.integration.management.cli.extensions.CliExtCommandsExtension
