+++ testsuite/standalone/src/test/resources/org/jboss/as/test/integration/management/cli/extensions/test-cli-ext-commands.txt
